finilize the calculator
