+++ src/Components/calculator.js
@@ -1,85 +1,20 @@
-import React, { useState } from 'react';
+// Calculator.js
+import React from 'react';
 import './calculator.css';
 import CalculatorScreen from './CalculatorScreen';
 import CalculatorButtons from './CalculatorButtons';
+import { useCalculatorLogic } from '../logic/calculate';
 
 function Calculator() {
-  const [inputValue, setInputValue] = useState('0');
-  const [previousValue, setPreviousValue] = useState(null);
-  const [operator, setOperator] = useState(null);
-
-  const handleNumberClick = (number) => {
-    if (inputValue === '0' || inputValue === 'Error') {
-      setInputValue(number.toString());
-    } else {
-      setInputValue(inputValue + number);
-    }
-  };
-
-  const handleOperatorClick = (op) => {
-    if (op === '%') {
-      handlePercentClick();
-    } else {
-      if (operator !== null && previousValue !== null) {
-        handleEqualClick();
-      }
-      setOperator(op);
-      setPreviousValue(inputValue);
-      setInputValue('0');
-    }
-  };
-
-  const handleEqualClick = () => {
-    if (previousValue === null || operator === null) return;
-
-    const num1 = parseFloat(previousValue);
-    const num2 = parseFloat(inputValue);
-
-    let result;
-    switch (operator) {
-      case '+':
-        result = num1 + num2;
-        break;
-      case '-':
-        result = num1 - num2;
-        break;
-      case '*':
-        result = num1 * num2;
-        break;
-      case '/':
-        if (num2 !== 0) {
-          result = num1 / num2;
-        } else {
-          setInputValue('Error');
-          return;
-        }
-        break;
-      default:
-        return;
-    }
-
-    setInputValue(result.toString());
-    setPreviousValue(null);
-    setOperator(null);
-  };
-
-  const handleClearClick = () => {
-    setInputValue('0');
-    setPreviousValue(null);
-    setOperator(null);
-  };
-
-  const handleDecimalClick = () => {
-    if (!inputValue.includes('.')) {
-      setInputValue(inputValue + '.');
-    }
-  };
-
-  const handlePercentClick = () => {
-    const num = parseFloat(inputValue);
-    const result = num / 100;
-    setInputValue(result.toString());
-  };
+  const {
+    inputValue,
+    handleNumberClick,
+    handleOperatorClick,
+    handleEqualClick,
+    handleClearClick,
+    handleToggleSign,
+    handleDecimalClick,
+  } = useCalculatorLogic();
 
   return (
     <div className="calculator">
@@ -89,6 +24,7 @@ function Calculator() {
         onOperatorClick={handleOperatorClick}
         onEqualClick={handleEqualClick}
         onClearClick={handleClearClick}
+        onToggleSign={handleToggleSign}
         onDecimalClick={handleDecimalClick}
       />
     </div>
